p2p: make NodeID and NetAddress public
